// server/src/gameLogic/Enums/AreaNames.ts
export enum AreaName {
    FORLINDON = 'FORLINDON',
    MITHLOND = 'MITHLOND',
    LUNE_VALLEY = 'LUNE_VALLEY',
    EVENDIM_HILLS = 'EVENDIM_HILLS',
    TOWER_HILLS = 'TOWER_HILLS',
    THE_SHIRE = 'THE_SHIRE',
    HARLINDON = 'HARLINDON',
    FORODWAITH = 'FORODWAITH',
    ANGMAR = 'ANGMAR',
    EASTERN_ANGMAR = 'EASTERN_ANGMAR',
    BORDERLANDS = 'BORDERLANDS',
    NORTH_DOWNS = 'NORTH_DOWNS',
    RHUDAUR = 'RHUDAUR',
    FORNOST = 'FORNOST',
    WEATHER_HILLS = 'WEATHER_HILLS',
    OLD_FOREST = 'OLD_FOREST',
    BUCKLAND = 'BUCKLAND',
    SOUTH_DOWNS = 'SOUTH_DOWNS',
    CARROCK = 'CARROCK',
    NORTH_MIRKWOOD = 'NORTH_MIRKWOOD',
    EASTERN_MIRKWOOD = 'EASTERN_MIRKWOOD',
    SOUTH_MIRKWOOD = 'SOUTH_MIRKWOOD',
    ANDUIN_VALLEY = 'ANDUIN_VALLEY',
    NORTH_RHUN = 'NORTH_RHUN',
    WITHERED_HEATH = 'WITHERED_HEATH',
    SOUTH_RHUN = 'SOUTH_RHUN',
    ESGAROTH = 'ESGAROTH',
    MINHIRIATH = 'MINHIRIATH',
    EREGION = 'EREGION',
    DUNLAND = 'DUNLAND',
    ENEDWAITH = 'ENEDWAITH',
    FANGORN = 'FANGORN',
    WEST_ROHAN = 'WEST_ROHAN',
    GAP_OF_ROHAN = 'GAP_OF_ROHAN',
    MORIA = 'MORIA',
    GLADDEN_FIELDS = 'GLADDEN_FIELDS',
    LORIEN = 'LORIEN',
    THE_WOLD = 'THE_WOLD',
    EMYN_MUIL = 'EMYN_MUIL',
    BROWNLANDS = 'BROWNLANDS',
    RHUN_HILLS = 'RHUN_HILLS',
    DEAD_MARSHES = 'DEAD_MARSHES',
    DRUWAITH_IAUR = 'DRUWAITH_IAUR', 
    ANDRAST = 'ANDRAST',
    ANFALAS = 'ANFALAS',
    VALE_OF_ERECH = 'VALE_OF_ERECH',
    LAMEDON = 'LAMEDON',
    BELFALAS = 'BELFALAS',
    LEBENNIN = 'LEBENNIN',
    MINAS_TIRITH = 'MINAS_TIRITH',
    ITHILIEN = 'ITHILIEN',
    SOUTH_ITHILIEN = 'SOUTH_ITHILIEN',
    UDUN_VALE = 'UDUN_VALE',
    MOUNT_DOOM = 'MOUNT_DOOM',
    BARAD_DUR = 'BARAD_DUR',
    MINAS_MORGUL = 'MINAS_MORGUL',
    GORGOROTH = 'GORGOROTH',
    NURN = 'NURN',
    HARONDOR = 'HARONDOR',
    UMBAR = 'UMBAR',
    DEEP_HARAD = 'DEEP_HARAD',
    HARAD = 'HARAD',
    NEAR_HARAD = 'NEAR_HARAD',
    KHAND = 'KHAND',
}